Use a left click instead of a right click to trigger the menu.
The goal is to add compatibility with OS that display the contextual menu when the mouse right button is pushed and not when it is released (on mouse down instead of on mouse click). Before that, there was no way for the extension to pop before the apparition of the contextual menu.